fix(capability): the poll timeout is a float, not an int
pyright caught this at 89 -> 91 after the previous commit: the deadline is float
arithmetic (`time.monotonic()`, `MCP_POLL_INTERVAL = 1.0`), so a caller asking
for 4.5s is legitimate and the `int` annotation was simply wrong. Widened rather
than changing the tests to pass whole seconds — a non-multiple timeout is exactly
what makes an off-by-one-interval bug visible.

Back to 89 = 89 against the baseline on 292ecee.

+++ src/harnessed/capability.py
@@ -674,7 +674,7 @@ def introspect_mcp(
     harness: str = "claude",
     *,
     expect: Collection[str] = (),
-    timeout: int = MCP_CONNECT_TIMEOUT,
+    timeout: float = MCP_CONNECT_TIMEOUT,
 ) -> tuple[dict[str, str], str]:
     """Return ({connected server -> status}, source-label), preferring machine-readable sources.
